+++ src/MixedModels.jl
@@ -211,7 +211,8 @@ using PrecompileTools
         fit(MixedModel,
             @formula(use ~ 1 + age + abs2(age) + urban + livch + (1 | urban & dist)),
             contra,
-            Bernoulli(); progress)
+            Bernoulli();
+            progress)
     end
 end