+++ src/bootstrap.jl
@@ -159,9 +159,25 @@ function Base.reduce(::typeof(vcat), v::AbstractVector{MixedModelBootstrap{T}}) 
         deepcopy(b1.fcnames))
 end
 
+function Base.show(io::IO, mime::MIME"text/plain", x::MixedModelBootstrap)
+    tbl = x.tbl
+    println(io, "MixedModelBootstrap with $(length(x)) samples")
+    out = NamedTuple[]
+    for col in Tables.columnnames(tbl)
+        col == :obj && continue
+        s = summarystats(Tables.getcolumn(tbl, col))
+        push!(out, (; parameter=col, s.min, s.q25, s.median, s.mean, s.q75, s.max))
+    end
+    tt = FlexTable(out)
+    # trim out the FlexTable header
+    str = last(split(sprint(show, mime, tt), "\n"; limit=2))
+    println(io, str)
+    return nothing
+end
+
 """
     parametricbootstrap([rng::AbstractRNG], nsamp::Integer, m::MixedModel{T}, ftype=T;
-        β = coef(m), σ = m.σ, θ = m.θ, hide_progress=false, optsum_overrides=(;))
+        β = coef(m), σ = m.σ, θ = m.θ, progress=true, optsum_overrides=(;))
 
 Perform `nsamp` parametric bootstrap replication fits of `m`, returning a `MixedModelBootstrap`.
 
@@ -177,10 +193,10 @@ performance benefits.
 - `β`, `σ`, and `θ` are the values of `m`'s parameters for simulating the responses.
 - `σ` is only valid for `LinearMixedModel` and `GeneralizedLinearMixedModel` for
 families with a dispersion parameter.
-- `hide_progress` can be used to disable the progress bar. Note that the progress
+- `progress` controls whehter the progress bar is shown. Note that the progress
 bar is automatically disabled for non-interactive (i.e. logging) contexts.
 - `optsum_overrides` is used to override values of [OptSummary](@ref) in the models
-fit during the bootstrapping process. For example, `optsum_overrides=(;ftol_rel=1e08)`
+fit during the bootstrapping process. For example, `optsum_overrides=(;ftol_rel=1e-08)`
 reduces the convergence criterion, which can greatly speed up the bootstrap fits.
 Taking advantage of this speed up to increase `n` can often lead to better estimates
 of coverage intervals.
@@ -194,9 +210,17 @@ function parametricbootstrap(
     σ=morig.σ,
     θ::AbstractVector=morig.θ,
     use_threads::Bool=false,
-    hide_progress::Bool=false,
+    progress::Bool=true,
+    hide_progress::Union{Bool,Nothing}=nothing,
     optsum_overrides=(;),
 ) where {T}
+    if !isnothing(hide_progress)
+        Base.depwarn(
+            "`hide_progress` is deprecated, please use `progress` instead." *
+            "NB: `progress` is a positive action, i.e. `progress=true` means show the progress bar.",
+            :parametricbootstrap; force=true)
+        progress = !hide_progress
+    end
     if σ !== missing
         σ = T(σ)
     end
@@ -216,7 +240,7 @@ function parametricbootstrap(
         "use_threads is deprecated and will be removed in a future release",
         :parametricbootstrap,
     )
-    samp = replicate(n; hide_progress=hide_progress) do
+    samp = replicate(n; progress) do
         simulate!(rng, m; β, σ, θ)
         refit!(m; progress=false)
         # @info "" m.optsum.feval
@@ -253,7 +277,7 @@ of `iter`, `type`, `group`, `name` and `value`.
     a breaking change.
 """
 function allpars(bsamp::MixedModelFitCollection{T}) where {T}
-    fits, λ, fcnames = bsamp.fits, bsamp.λ, bsamp.fcnames
+    (; fits, λ, fcnames) = bsamp
     npars = 2 + length(first(fits).β) + sum(map(k -> (k * (k + 1)) >> 1, size.(bsamp.λ, 2)))
     nresrow = length(fits) * npars
     cols = (
@@ -303,9 +327,13 @@ function allpars(bsamp::MixedModelFitCollection{T}) where {T}
 end
 
 """
-    confint(pr::MixedModelBootstrap; level::Real=0.95)
+    confint(pr::MixedModelBootstrap; level::Real=0.95, method=:shortest)
 
 Compute bootstrap confidence intervals for coefficients and variance components, with confidence level level (by default 95%).
+
+The keyword argument `method` determines whether the `:shortest`, i.e. highest density, interval is used 
+or the `:equaltail`, i.e. quantile-based, interval is used. For historical reasons, the default is `:shortest`, 
+but `:equaltail` gives the interval that is most comparable to the profile and Wald confidence intervals.
 
 !!! note
     The API guarantee is for a Tables.jl compatible table. The exact return type is an
@@ -319,7 +347,11 @@ Compute bootstrap confidence intervals for coefficients and variance components,
 
 See also [`shortestcovint`](@ref).
 """
-function StatsBase.confint(bsamp::MixedModelBootstrap{T}; level::Real=0.95) where {T}
+function StatsBase.confint(
+    bsamp::MixedModelBootstrap{T}; level::Real=0.95, method=:shortest
+) where {T}
+    method in [:shortest, :equaltail] ||
+        throw(ArgumentError("`method` must be either :shortest or :equaltail."))
     cutoff = sqrt(quantile(Chisq(1), level))
     # Creating the table is somewhat wasteful because columns are created then immediately skipped.
     tbl = Table(bsamp.tbl)
@@ -333,8 +365,13 @@ function StatsBase.confint(bsamp::MixedModelBootstrap{T}; level::Real=0.95) wher
             ),
         ),
     )
+    tails = [(1 - level) / 2, (1 + level) / 2]
     for p in par
-        l, u = shortestcovint(sort!(copyto!(v, getproperty(tbl, p))), level)
+        if method === :shortest
+            l, u = shortestcovint(sort!(copyto!(v, getproperty(tbl, p))), level)
+        else
+            l, u = quantile(getproperty(tbl, p), tails)
+        end
         push!(lower, l)
         push!(upper, u)
     end
@@ -392,12 +429,12 @@ function Base.propertynames(bsamp::MixedModelFitCollection)
 end
 
 """
+    setθ!(bsamp::MixedModelFitCollection, θ::AbstractVector)
     setθ!(bsamp::MixedModelFitCollection, i::Integer)
 
 Install the values of the i'th θ value of `bsamp.fits` in `bsamp.λ`
 """
-function setθ!(bsamp::MixedModelFitCollection, i::Integer)
-    θ = bsamp.fits[i].θ
+function setθ!(bsamp::MixedModelFitCollection{T}, θ::AbstractVector{T}) where {T}
     offset = 0
     for (λ, inds) in zip(bsamp.λ, bsamp.inds)
         λdat = _getdata(λ)
@@ -408,6 +445,10 @@ function setθ!(bsamp::MixedModelFitCollection, i::Integer)
         offset += length(inds)
     end
     return bsamp
+end
+
+function setθ!(bsamp::MixedModelFitCollection, i::Integer)
+    return setθ!(bsamp, bsamp.θ[i])
 end
 
 _getdata(x::Diagonal) = x
@@ -444,7 +485,7 @@ Return the shortest interval containing `level` proportion for each parameter fr
     a breaking change.
 """
 function shortestcovint(bsamp::MixedModelFitCollection{T}, level=0.95) where {T}
-    allpars = bsamp.allpars
+    allpars = bsamp.allpars  # TODO probably simpler to use .tbl instead of .allpars
     pars = unique(zip(allpars.type, allpars.group, allpars.names))
 
     colnms = (:type, :group, :names, :lower, :upper)
@@ -521,6 +562,7 @@ end
 
 """
     tidyσs(bsamp::MixedModelFitCollection)
+
 Return a tidy (row)table with the estimates of the variance components (on the standard deviation scale) spread into columns
 of `iter`, `group`, `column` and `σ`.
 """
@@ -544,40 +586,90 @@ function tidyσs(bsamp::MixedModelFitCollection{T}) where {T}
     return result
 end
 
-function pbstrtbl(bsamp::MixedModelFitCollection{T}) where {T}
-    (; fits, λ, inds) = bsamp
-    row1 = first(fits)
-    cnms = [:obj, :σ]
-    pos = Dict{Symbol,UnitRange{Int}}(:obj => 1:1, :σ => 2:2)
-    βsz = length(row1.β)
-    append!(cnms, _generatesyms('β', βsz))
-    lastpos = 2 + βsz
-    pos[:β] = 3:lastpos
-    σsz = sum(m -> size(m, 1), bsamp.λ)
-    append!(cnms, _generatesyms('σ', σsz))
-    pos[:σs] = (lastpos + 1):(lastpos + σsz)
-    lastpos += σsz
-    θsz = length(row1.θ)
-    append!(cnms, _generatesyms('θ', θsz))
-    pos[:θ] = (lastpos + 1):(lastpos + θsz)
-    tblrowtyp = NamedTuple{(cnms...,),NTuple{length(cnms),T}}
-    val = sizehint!(tblrowtyp[], length(bsamp.fits))
-    v = Vector{T}(undef, length(cnms))
-    for (i, r) in enumerate(bsamp.fits)
-        v[1] = r.objective
-        v[2] = coalesce(r.σ, one(T))
-        copyto!(view(v, pos[:β]), r.β)
-        fill!(view(v, pos[:σs]), zero(T))
-        copyto!(view(v, pos[:θ]), r.θ)
-        setθ!(bsamp, i)
-        vpos = first(pos[:σs])
-        for l in λ
-            for λr in eachrow(l)
-                v[vpos] = r.σ * norm(λr)
-                vpos += 1
-            end
+_nρ(d::Diagonal) = 0
+_nρ(t::LowerTriangular) = kchoose2(size(t.data, 1))
+
+function σρnms(λ)
+    σsyms = _generatesyms('σ', sum(first ∘ size, λ))
+    ρsyms = _generatesyms('ρ', sum(_nρ, λ))
+    val = sizehint!(Symbol[], length(σsyms) + length(ρsyms))
+    for l in λ
+        for _ in axes(l, 1)
+            push!(val, popfirst!(σsyms))
         end
-        push!(val, tblrowtyp(v))
+        for _ in 1:_nρ(l)
+            push!(val, popfirst!(ρsyms))
+        end
     end
     return val
+end
+
+function _syms(bsamp::MixedModelBootstrap)
+    (; fits, λ) = bsamp
+    (; β, θ) = first(fits)
+    syms = [:obj]
+    append!(syms, _generatesyms('β', length(β)))
+    push!(syms, :σ)
+    append!(syms, σρnms(λ))
+    return append!(syms, _generatesyms('θ', length(θ)))
+end
+
+function σρ!(v::AbstractVector, d::Diagonal, σ)
+    return append!(v, σ .* d.diag)
+end
+
+function σρ!(v::AbstractVector{T}, t::LowerTriangular{T}, σ::T) where {T}
+    """
+        σρ!(v, t, σ)
+    push! `σ` times the row lengths (σs) and the inner products of normalized rows (ρs) of `t` onto `v` 
+    """
+    dat = t.data
+    for i in axes(dat, 1)
+        ssqr = zero(T)
+        for j in 1:i
+            ssqr += abs2(dat[i, j])
+        end
+        len = sqrt(ssqr)
+        push!(v, σ * len)
+        if len > 0
+            for j in 1:i
+                dat[i, j] /= len
+            end
+        end
+    end
+    for i in axes(dat, 1)
+        for j in 1:(i - 1)
+            s = zero(T)
+            for k in 1:i
+                s += dat[i, k] * dat[j, k]
+            end
+            push!(v, s)
+        end
+    end
+    return v
+end
+
+function pbstrtbl(bsamp::MixedModelFitCollection{T}) where {T}
+    (; fits, λ) = bsamp
+    λcp = copy.(λ)
+    syms = _syms(bsamp)
+    m = length(syms)
+    n = length(fits)
+    v = sizehint!(T[], m * n)
+    for f in fits
+        (; β, θ, σ) = f
+        push!(v, f.objective)
+        append!(v, β)
+        push!(v, σ)
+        setθ!(bsamp, θ)
+        for l in λ
+            σρ!(v, l, σ)
+        end
+        append!(v, θ)
+    end
+    m = permutedims(reshape(v, (m, n)), (2, 1)) # equivalent to collect(transpose(...))
+    for k in eachindex(λ, λcp)   # restore original contents of λ
+        copyto!(λ[k], λcp[k])
+    end
+    return Table(Tables.table(m; header=syms))
 end